assignments/module17/q5/serializers.py
from rest_framework import serializers
from q3.models import Doctor


class DoctorSerializer(serializers.ModelSerializer):
    class Meta:
        model = Doctor
        fields = '__all__'
        read_only_fields = ['id', 'created_at']
