assignments/module17/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.6
Django==6.0.3
django-allauth==65.15.1
django-environ==0.13.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
frozenlist==1.8.0
idna==3.11
MarkupSafe==3.0.3
multidict==6.7.1
pillow==12.2.0
propcache==0.4.1
pycparser==3.0
PyJWT==2.12.1
python-dotenv==1.2.2
python-http-client==3.3.7
requests==2.33.1
sendgrid==6.12.5
sqlparse==0.5.5
stripe==15.0.1
twilio==9.10.4
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.8
yarl==1.23.0
